1.5.1 Exercises. 2: toCelsius

+++ Chap01/1_5_1Exercises/1/factorial.cpp
@@ -1,9 +1,9 @@
 #include <iostream>
 
-long factorial(long n)
+long long factorial(long n)
 {
-    long ans(1);
-    for (long i=2; i<=n; ++i ){
+    long long ans(1);
+    for (long long i=2; i<=n; ++i ){
         ans=ans*i;
 //        std::cout<<ans<<std::endl;
     }
