+++ Chap01/1_5_1Exercises/1/main.cpp
@@ -7,11 +7,11 @@ int main()//int argc, char *argv[])
 //    QCoreApplication a(argc, argv);
     using namespace std;
 
-    long max(1);//max valid number
+    long long max(1);//max valid number
 
-    for (long n=2; ; ++n){
-        long nfact=factorial(n);
-        long m=n;
+    for (long long n=2; ; ++n){
+        long long nfact=factorial(n);
+        long long m=n;
         do{
             nfact=nfact/m;
             --m;
